+++ mpire/signal.py
@@ -1,21 +1,18 @@
 from inspect import Traceback
 from signal import getsignal, SIG_IGN, SIGINT, signal as signal_, Signals
+from threading import current_thread, main_thread
 from types import FrameType
 from typing import Type
 
 
 class DelayedKeyboardInterrupt:
 
-    def __init__(self, in_thread: bool = False) -> None:
-        """
-        :param in_thread: Whether or not we're living in a thread or not
-        """
-        self.in_thread = in_thread
+    def __init__(self) -> None:
         self.signal_received = None
 
     def __enter__(self) -> None:
         # When we're in a thread we can't use signal handling
-        if not self.in_thread:
+        if current_thread() == main_thread():
             self.signal_received = False
             self.old_handler = signal_(SIGINT, self.handler)
 
@@ -23,7 +20,7 @@ class DelayedKeyboardInterrupt:
         self.signal_received = (sig, frame)
 
     def __exit__(self, exc_type: Type, exc_val: Exception, exc_tb: Traceback) -> None:
-        if not self.in_thread:
+        if current_thread() == main_thread():
             signal_(SIGINT, self.old_handler)
             if self.signal_received:
                 self.old_handler(*self.signal_received)
@@ -32,13 +29,15 @@ class DelayedKeyboardInterrupt:
 class DisableKeyboardInterruptSignal:
 
     def __enter__(self) -> None:
-        # Prevent signal from propagating to child process
-        self._handler = getsignal(SIGINT)
-        ignore_keyboard_interrupt()
+        if current_thread() == main_thread():
+            # Prevent signal from propagating to child process
+            self._handler = getsignal(SIGINT)
+            ignore_keyboard_interrupt()
 
     def __exit__(self, exc_type: Type, exc_val: Exception, exc_tb: Traceback) -> None:
-        # Restore signal
-        signal_(SIGINT, self._handler)
+        if current_thread() == main_thread():
+            # Restore signal
+            signal_(SIGINT, self._handler)
 
 
 def ignore_keyboard_interrupt():