+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+Dev
+---
+
+* A WorkerPool can now be started within a thread which isn't the main thread (`#44`_)
+
+.. _#44: https://github.com/Slimmer-AI/mpire/issues/44
+
+
 2.3.5
 -----